*** empty log message ***

+++ magick/defines.h
@@ -144,8 +144,8 @@ extern "C" {
      (((double) (color).blue-(double) (target).blue)* \
       ((double) (color).blue-(double) (target).blue))) <= \
       (double) (distance*distance))
-#define CompositeOver(P,Pa,Q,Qa)  ((((double) (Pa)*(P)+(Qa)*((Q)- \
-  (double) (Pa)*(P)/MaxRGB))/((Pa)+(Qa)-(double) (Pa)*(Qa)/MaxRGB)))
+#define CompositeOver(P,Pa,Q,Qa)  (((double) (Pa)*(P)+(double) (Qa)*((Q)- \
+  (double) (Pa)*(P)/MaxRGB))/((Pa)+(Qa)-(double) (Pa)*(Qa)/MaxRGB))
 #define Extent(string)  ((int) strlen(string))
 #define False  0
 #define DegreesToRadians(x) (MagickPI*(x)/180.0)
